Test refactorings for improved code coverage
Added .runsettings file for configuring code coverage, separated OpenXmlImporter and OpenXmlExporter tests, added new tests for retrieving sheet informations, sheet dimensions and column names, added new queryrange tests, added new addpicture test, removed obsolete FileHelper test utility class
