+++ tests/MiniExcel.OpenXml.Tests/AlterSheets/MiniExcelAlterSheetsTests.cs
@@ -1,4 +1,5 @@
 using MiniExcelLib.OpenXml.Models;
+using MiniExcelLib.Tests.Common.Utils;
 using static MiniExcelLib.OpenXml.Tests.Utils.SheetHelper;
 
 namespace MiniExcelLib.OpenXml.Tests.AlterSheets;
@@ -98,18 +99,23 @@ public class MiniExcelAlterSheetTests
     }
 
     [Fact]
-    public void AlterSheet_WhenNoOptionalParametersProvided_LeavesSheetUnchanged()
+    public async Task AlterSheet_WhenNoOptionalParametersProvided_LeavesSheetUnchanged()
     {
         // Arrange
         const string targetSheet = "Sheet1";
-        using var stream = CreateTestWorkbookStream();
+        using var path = AutoDeletingPath.Create();
+        using (var stream = CreateTestWorkbookStream())
+        {
+            using var fs = File.OpenWrite(path.FilePath);
+            stream.Position = 0;
+            stream.CopyTo(fs);
+        }
 
         // Act
-        _excelExporter.AlterSheet(stream, targetSheet);
+        _excelExporter.AlterSheet(path.FilePath, targetSheet);
 
         // Assert
-        stream.Position = 0;
-        using var package = new ExcelPackage(stream);
+        using var package = new ExcelPackage(path.FilePath);
         var sheet = package.Workbook.Worksheets[targetSheet];
 
         // Ensure defaults remain intact
