+++ tests/MiniExcel.OpenXml.Tests/Templates/MiniExcelTemplateAsyncTests.cs
@@ -1,4 +1,6 @@
-﻿using MiniExcelLib.OpenXml.Tests.Utils;
+﻿using ClosedXML.Excel;
+using MiniExcelLib.OpenXml.Picture;
+using MiniExcelLib.OpenXml.Tests.Utils;
 using MiniExcelLib.Tests.Common.Utils;
 
 namespace MiniExcelLib.OpenXml.Tests.SaveByTemplate;
@@ -8,6 +10,11 @@ public class MiniExcelTemplateAsyncTests
     private readonly OpenXmlImporter _excelImporter =  MiniExcel.Importers.GetOpenXmlImporter();
     private readonly OpenXmlTemplater _excelTemplater =  MiniExcel.Templaters.GetOpenXmlTemplater();
     
+    static MiniExcelTemplateAsyncTests()
+    {
+        ExcelPackage.LicenseContext = LicenseContext.NonCommercial;
+    }
+
     [Fact]
     public async Task DatatableTemptyRowTest()
     {
@@ -832,7 +839,12 @@ public class MiniExcelTemplateAsyncTests
                     new { name = "Keaton", department = "IT" }
                 }
             };
-            await _excelTemplater.FillTemplateAsync(path.ToString(), templatePath, value);
+
+            await using (var stream = File.OpenWrite(path.FilePath))
+            {
+                await using var templateStream = File.OpenRead(templatePath);
+                await _excelTemplater.FillTemplateAsync(stream, templatePath, value);
+            }
 
             var rows = await _excelImporter.QueryAsync(path.ToString()).ToListAsync();
             Assert.Equal("FooCompany", rows[0].A);
@@ -878,7 +890,6 @@ public class MiniExcelTemplateAsyncTests
         });
     }
     
-    
     [Fact]
     public async Task TestMergeSameCellsWithTagAsync()
     {
@@ -898,8 +909,9 @@ public class MiniExcelTemplateAsyncTests
     {
         var path = PathHelper.GetFile("xlsx/TestMergeWithLimitTag.xlsx");
         using var mergedFilePath = AutoDeletingPath.Create();
+        await using var stream = new FileStream(mergedFilePath.FilePath, FileMode.OpenOrCreate, FileAccess.ReadWrite);
 
-        await _excelTemplater.MergeSameCellsAsync(mergedFilePath.ToString(), path);
+        await _excelTemplater.MergeSameCellsAsync(stream, path);
         var mergedCells = SheetHelper.GetFirstSheetMergedCells(mergedFilePath.ToString());
 
         Assert.Equal("A3:A4", mergedCells[0]);
@@ -912,6 +924,7 @@ public class MiniExcelTemplateAsyncTests
     public async Task TestIEnumerableWithFormulas()
     {
         var templatePath = PathHelper.GetFile("xlsx/TestTemplateBasicIEnumerableFillWithFormulas.xlsx");
+        await using var templateStream = File.OpenRead(templatePath);
         using var path = AutoDeletingPath.Create();
 
         var value = new
@@ -926,9 +939,26 @@ public class MiniExcelTemplateAsyncTests
                 new { name = "Joan", department = "IT", salary = 120000 }
             }
         };
-        await _excelTemplater.FillTemplateAsync(path.FilePath, templatePath, value, true);
+        await _excelTemplater.FillTemplateAsync(path.FilePath, templateStream, value, true);
 
         var dimension = SheetHelper.GetFirstSheetDimensionRefValue(path.FilePath);
         Assert.Equal("A1:C13", dimension);
+    }
+
+    [Fact]
+    public async Task AddPictureTest()
+    {
+        MiniExcelPicture[] pictures = 
+        [
+            new() { ImageBytes = File.ReadAllBytes(PathHelper.GetFile("images/github_logo.png")), CellAddress = "B2", HeightPx = 40, WidthPx = 40 },
+            new() { ImageBytes = File.ReadAllBytes(PathHelper.GetFile("images/google_logo.png")), CellAddress = "C3", HeightPx = 45, WidthPx = 45 },
+            new() { ImageBytes = File.ReadAllBytes(PathHelper.GetFile("images/microsoft_logo.png")), CellAddress = "D4", HeightPx = 50, WidthPx = 50 },
+            new() { ImageBytes = File.ReadAllBytes(PathHelper.GetFile("images/reddit_logo.png")), CellAddress = "E5", HeightPx = 55, WidthPx = 55 },
+        ];
+
+        using var path = AutoDeletingPath.Create();
+        await MiniExcel.Exporters.GetOpenXmlExporter().ExportAsync(path.ToString(), Array.Empty<Dictionary<string, object>>());
+
+        await _excelTemplater.AddPictureAsync(path.ToString(), images: pictures);
     }
 }
