+++ tests/MiniExcel.OpenXml.Tests/Utils/SheetHelper.cs
